Create non-io constants file and reduce file, move logic there
